Only 1 entry point
